chore: 서버액션을 사용하는 함수 통합
기존의 api/auth.ts에서 사용하던 함수들을 제거하고 /actions/auth.ts로 통합했습니다.
